Add pipeline generation and policy data
I created a stripped down version of the docker-build-oci-ta pipeline
without some of the unnecessary scans and with the verify-source task.
This is in the managed context because a privileged user would need to
run this in order to generate the acceptable task bundles image.

I also created a policy data directory that we can use for customizing
our own policies. This requires all of the tasks in our pipeline except
for the image index task.

Updated the TODOs to mark these items as needing verification.

Signed-off-by: arewm <arewm@users.noreply.github.com>